requirements.txt
snowflake-snowpark-python
snowflake-ml-python
pandas
streamlit